requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
Deprecated==1.2.14
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.13.4
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.3.1
ftfy==6.2.0
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
kubernetes==29.0.0
lxml==5.2.1
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mlx==0.11.1
mlx-clip @ git+https://github.com/harperreed/mlx_clip.git@f56e3ecc72c74c68b6b50eb6f50c3f22fc23fe2c
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
open-clip-torch==2.24.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==24.0
parsel==1.9.1
pillow==10.3.0
pluggy==1.5.0
posthog==3.5.0
Protego==0.3.1
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
queuelib==1.6.2
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
Scrapy==2.11.1
sentencepiece==0.2.0
service-identity==24.1.0
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
timm==0.9.16
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.1
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
zope.interface==6.3
